Rename StaticStabilityGravity into ConvexShapeMatcher
