+++ CMakeLists.txt
@@ -33,6 +33,7 @@ SET(${PROJECT_NAME}_HEADERS
   include/hpp/constraints/fwd.hh
   include/hpp/constraints/tools.hh
   include/hpp/constraints/convex-shape.hh
+  include/hpp/constraints/convex-shape-matcher.hh
   include/hpp/constraints/symbolic-function.hh
   include/hpp/constraints/orientation.hh
   include/hpp/constraints/position.hh
